requirements.txt
flask==3.0.2
redis==5.0.2
fakeredis==2.21.1
bcrypt==4.1.2
pyjwt==2.8.0
boto3==1.34.50
prometheus-client==0.20.0
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-instrumentation-flask==0.44b0
opentelemetry-exporter-otlp==1.23.0
gunicorn==21.2.0
pytest==8.0.2
pytest-cov==4.1.0
pytest-benchmark==4.0.0
schemathesis==3.25.4
hypothesis==6.98.15
moto==5.0.2
tenacity==8.2.3
SQLAlchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
structlog==24.1.0
cryptography==42.0.5
